requirements.txt
# Core dependencies
crewai>=0.150.0
crewai-tools>=0.59.0
stagehand-py==0.3.9

# Data handling
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Utilities
python-dotenv>=1.0.0
rich>=13.0.0
typer>=0.9.0

# Optional: For local development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.0.0
isort>=5.12.0